requirements.txt
# Vercel Python 依赖
beautifulsoup4==4.13.5
Flask==3.1.2
Flask-CORS==4.0.1
requests==2.32.5
SQLAlchemy==2.0.43
openai>=1.0.0
psycopg2-binary
python-dotenv>=1.0.0
supabase>=2.0.0
